LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-units-invalidation-expected.txt

Test

FAIL cqi respond when selected container changes type (inline-size -> none) assert_equals: expected "50px" but got "30px"
FAIL cqb respond when selected container changes type (size -> none) assert_equals: expected "60px" but got "40px"
FAIL cqb respond when intermediate container changes type (inline-size -> size) assert_equals: expected "20px" but got "40px"
FAIL cqi respond when selected container changes inline-size assert_equals: expected "5px" but got "30px"
FAIL cqb respond when selected container changes block-size assert_equals: expected "5px" but got "40px"
